Do not use index_suppressed if starting rd DMA is slow
Seeking to a new track may be slow because of writes that are draining
(due to async I/O). And filling dma_rd may be slow simply because of a
random slow read from flash. When index_suppressed == FALSE, it is be
better to further delay RDATA than to change the index timing.