Optimize write_bc size for async
The large write_bc needed with blocking I/O serves little purpose with
all images using asynchronous I/O. This increases the number of images
that can fit the full track in memory.

Tracks that fit fully in memory can buffer the reads and writes in the
image without the help of write_bc. And tracks that don't fit in memory
benefit from smaller read_bc, beside the extra buffer space, because the
image processing runs "ahead" of the read/write head. When a write
starts, that goes "back in time" from the image's perspective, so the
image needs to keep that data around which duplicates read_bc or
potentially re-read it. Reducing the size of read_bc gives the image
more memory to buffer and reduces the amount of needed data that has to
be kept around or re-read for writes.