da: Convert to async I/O
We use a custom write buffer for "open loop" writes. Reads do not
_block_ on the write buffer but are delayed until it is drained. This
lets us use all async code paths while also supporting lots of writes
without the controller doing any reads or index synchronization.

The read caching is preserved when writing. This is partly because the
default track size is small enough (4KB) that we can just pay the cost
all the time. But it also avoids the need to wait for a pending read to
complete to make memory available for processing written commands.
However, if the D-A controller extends the track size past 8 sectors
with CMD_SET_LBA, then the read cache becomes useless.